requirements.txt
streamlit
sentence-transformers
spacy
scikit-learn
numpy
pdfplumber
python-docx
rapidfuzz
lime
shap
pandas
matplotlib
